#29 refactor-resources의 fxml user, admin으로 나누기
