--- src/main/java/com/sysone/ogamza/controller/UserHomeController.java
@@ -1,17 +1,0 @@
-package com.sysone.ogamza.controller;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-import javafx.scene.effect.DropShadow;
-import javafx.scene.image.ImageView;
-import javafx.scene.paint.Color;
-
-public class UserHomeController {
-
-
-    public void initialize() {
-
-
-    }
-
-}
