+++ src/main/java/com/sysone/ogamza/dao/user/EmojiDAO.java
