+++ src/main/java/com/sysone/ogamza/dao/user/TodayFortuneDAO.java
