+++ src/main/java/com/sysone/ogamza/dto/user/ScheduleContentDTO.java
