+++ src/main/java/com/sysone/ogamza/dto/user/ScheduleListDTO.java
