+++ src/main/java/com/sysone/ogamza/service/user/ScheduleService.java
@@ -2,7 +2,7 @@ package com.sysone.ogamza.service;
 
 import com.sysone.ogamza.dto.ScheduleContentDto;
 import com.sysone.ogamza.dto.ScheduleListDto;
-import com.sysone.ogamza.repository.ScheduleDAO;
+import com.sysone.ogamza.dao.user.ScheduleDAO;
 import lombok.Getter;
 
 import java.util.ArrayList;
