+++ src/main/resources/fxml/admin/AdminHome.fxml
@@ -33,7 +33,7 @@
                <children>
                   <ImageView fitHeight="150.0" fitWidth="200.0" pickOnBounds="true" preserveRatio="true" translateX="70.0" translateY="40.0">
                      <image>
-                        <Image url="@../images/employee.png" />
+                        <Image url="@../../images/employee.png" />
                      </image>
                   </ImageView>
                </children>
@@ -104,7 +104,7 @@
          <children>
             <ImageView fitHeight="40.0" fitWidth="40.0" pickOnBounds="true" preserveRatio="true">
                <image>
-                  <Image url="@../images/crown.png" />
+                  <Image url="@../../images/crown.png" />
                </image>
             </ImageView>
             <Text strokeType="OUTSIDE" strokeWidth="0.0" style="-fx-font-size: 26; -fx-text-alignment: center;" text="근태 랭킹" />
@@ -126,13 +126,13 @@
                      <children>
                         <ImageView fitHeight="25.0" fitWidth="25.0" pickOnBounds="true" preserveRatio="true">
                            <image>
-                              <Image url="@../images/left-quote.png" />
+                              <Image url="@../../images/left-quote.png" />
                            </image>
                         </ImageView>
                         <Text strokeType="OUTSIDE" strokeWidth="0.0" style="-fx-font-size: 28; -fx-font-family: 'ylee MHIM'; -fx-text-alignment: center;" text="잠올땐 자라" />
                         <ImageView fitHeight="25.0" fitWidth="25.0" pickOnBounds="true" preserveRatio="true">
                            <image>
-                              <Image url="@../images/right-quote.png" />
+                              <Image url="@../../images/right-quote.png" />
                            </image>
                         </ImageView>
                      </children>
