+++ src/main/resources/fxml/admin/AdminMainLayout.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane id="anchorPane" fx:id="anchorPane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="832.0" prefWidth="1284.0" style="-fx-background-radius: 50; -fx-background-color: #F3F5F9;" xmlns="http://javafx.com/javafx/24.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sysone.ogamza.controller.AdminMainController">
+<AnchorPane id="anchorPane" fx:id="anchorPane" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="832.0" prefWidth="1284.0" style="-fx-background-radius: 50; -fx-background-color: #F3F5F9;" xmlns="http://javafx.com/javafx/24.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sysone.ogamza.controller.admin.AdminMainController">
    <children>
       <VBox prefHeight="834.0" prefWidth="130.0" style="-fx-background-color: #7F90DB;">
          <children>
@@ -16,7 +16,7 @@
                <children>
                   <ImageView fitHeight="40.0" fitWidth="40.0" onMouseClicked="#goHome" pickOnBounds="true" preserveRatio="true">
                      <image>
-                        <Image url="@../images/home.png" />
+                        <Image url="@../../images/home.png" />
                      </image>
                   </ImageView>
                   <Label contentDisplay="CENTER" style="-fx-text-fill: #ffffff;" text="Home" textFill="WHITE">
@@ -29,7 +29,7 @@
                <children>
                   <ImageView fitHeight="40.0" fitWidth="40.0" onMouseClicked="#goDashboard" pickOnBounds="true" preserveRatio="true">
                      <image>
-                        <Image url="@../images/dashboard.png" />
+                        <Image url="@../../images/dashboard.png" />
                      </image>
                   </ImageView>
                   <Label alignment="CENTER" contentDisplay="CENTER" style="-fx-text-fill: #ffffff;" text="Dashboard" textFill="WHITE">
@@ -43,7 +43,7 @@
                <children>
                   <ImageView fitHeight="40.0" fitWidth="40.0" onMouseClicked="#goRecord" pickOnBounds="true" preserveRatio="true">
                      <image>
-                        <Image url="@../images/record.png" />
+                        <Image url="@../../images/record.png" />
                      </image>
                   </ImageView>
                   <Label style="-fx-text-fill: #ffffff;" text="Record" textFill="WHITE">
@@ -57,7 +57,7 @@
                <children>
                   <ImageView fitHeight="40.0" fitWidth="40.0" onMouseClicked="#goSettings" pickOnBounds="true" preserveRatio="true">
                      <image>
-                        <Image url="@../images/setting.png" />
+                        <Image url="@../../images/setting.png" />
                      </image>
                   </ImageView>
                   <Label style="-fx-text-fill: #ffffff;" text="Setting" textFill="WHITE">
@@ -71,7 +71,7 @@
                <children>
                   <ImageView fitHeight="40.0" fitWidth="40.0" onMouseClicked="#logout" pickOnBounds="true" preserveRatio="true">
                      <image>
-                        <Image url="@../images/logout.png" />
+                        <Image url="@../../images/logout.png" />
                      </image>
                   </ImageView>
                </children>
@@ -83,7 +83,7 @@
                <children>
                   <ImageView fitHeight="150.0" fitWidth="180.0" onMouseClicked="#goHome" pickOnBounds="true" preserveRatio="true">
                      <image>
-                        <Image url="@../images/userLogo.png" />
+                        <Image url="@../../images/userLogo.png" />
                      </image>
                   </ImageView>
                </children>
