+++ src/main/resources/fxml/user/Dashboard.fxml
@@ -13,7 +13,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane prefHeight="728.0" prefWidth="1155.0" stylesheets="@../css/dashboard.css" xmlns="http://javafx.com/javafx/24.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sysone.ogamza.controller.DashboardController">
+<AnchorPane prefHeight="728.0" prefWidth="1155.0" stylesheets="@../../css/dashboard.css" xmlns="http://javafx.com/javafx/24.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sysone.ogamza.controller.user.DashboardController">
    <children>
       <Label layoutX="58.0" layoutY="28.0" prefHeight="33.0" prefWidth="83.0" text="Day">
          <font>
