+++ src/main/resources/fxml/user/EditMemberModal.fxml
@@ -10,7 +10,7 @@
 
 <AnchorPane xmlns="http://javafx.com/javafx"
             xmlns:fx="http://javafx.com/fxml"
-            fx:controller="com.sysone.ogamza.controller.EditMemberModalController"
+            fx:controller="com.sysone.ogamza.controller.user.EditMemberModalController"
             prefHeight="400.0" prefWidth="600.0"
             style="-fx-background-color: white; -fx-background-radius:20;">
