+++ src/main/resources/fxml/user/ScheduleContent.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane prefHeight="431.0" prefWidth="360.0" style="-fx-background-color: #ffffff;" xmlns="http://javafx.com/javafx/24.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sysone.ogamza.controller.ScheduleContentController">
+<AnchorPane prefHeight="431.0" prefWidth="360.0" style="-fx-background-color: #ffffff;" xmlns="http://javafx.com/javafx/24.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sysone.ogamza.controller.user.ScheduleContentController">
     <VBox layoutX="4.0" layoutY="3.0" prefHeight="446.0" prefWidth="352.0" style="-fx-background-color: F4F9FF; -fx-background-radius: 30;">
         <children>
             <VBox prefHeight="70.0" prefWidth="352.0" style="-fx-alignment: top-center; -fx-padding: 30 0 0 0 0;">
