+++ src/main/resources/fxml/user/ScheduleList.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane prefHeight="449.0" prefWidth="570.0" style="-fx-background-color: #ffffff;" xmlns="http://javafx.com/javafx/24.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sysone.ogamza.controller.ScheduleListController">
+<AnchorPane prefHeight="449.0" prefWidth="570.0" style="-fx-background-color: #ffffff;" xmlns="http://javafx.com/javafx/24.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sysone.ogamza.controller.user.ScheduleListController">
     <VBox layoutX="4.0" layoutY="3.0" prefHeight="446.0" prefWidth="562.0" style="-fx-background-color: F4F9FF; -fx-background-radius: 30;">
         <children>
             <VBox prefHeight="87.0" prefWidth="492.0" style="-fx-alignment: top-center; -fx-padding: 30 0 20 0;">
