+++ src/main/resources/fxml/user/ScheduleRegister.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
-<AnchorPane prefHeight="431.0" prefWidth="360.0" style="-fx-background-color: #ffffff;" stylesheets="@../css/scheduleRegister.css" xmlns="http://javafx.com/javafx/24.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sysone.ogamza.controller.ScheduleRegisterController">
+<AnchorPane prefHeight="431.0" prefWidth="360.0" style="-fx-background-color: #ffffff;" stylesheets="@../../css/scheduleRegister.css" xmlns="http://javafx.com/javafx/24.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.sysone.ogamza.controller.user.ScheduleRegisterController">
    <VBox layoutX="4.0" layoutY="3.0" prefHeight="466.0" prefWidth="352.0" style="-fx-background-color: F4F9FF; -fx-background-radius: 30;">
       <children>
          <VBox prefHeight="70.0" prefWidth="352.0" style="-fx-alignment: top-center; -fx-padding: 30 0 0 0 0;">
